added independent class TEdge
