+++ AlibekovMR/02_lab_dheap/include/TGraph.h
@@ -2,28 +2,10 @@
 #define __TGRAPH__H__
 
 #include "TDisjointSet.h"
+#include "TEdge.h"
 #include "iostream"
 #include <queue>
 #include <utility> //temporarily (for pair)
-
-struct TEdge
-{
-  int startVertex;
-  int endVertex;
-  int weight;
-
-  TEdge() :startVertex(0), endVertex(0), weight(0) {};
-  TEdge(int _startVertex, int _endVertex, int _weight)
-    : startVertex(_startVertex), endVertex(_endVertex), weight(_weight) {};
-  TEdge(const TEdge& _edge);
-  ~TEdge() {};
-
-  bool operator>(const TEdge& _edge);
-  bool operator<(const TEdge& _edge);
-  bool operator==(const TEdge& _edge);
-
-  const TEdge& operator=(const TEdge& _edge);
-};
 
 class TGraph
 {
