+++ AlibekovMR/02_lab_dheap/sln/Heap/Heap.vcxproj
@@ -119,6 +119,7 @@
     <ClCompile Include="..\..\src\Algorithms.cpp" />
     <ClCompile Include="..\..\src\MyExceptions.cpp" />
     <ClCompile Include="..\..\src\TDisjointSet.cpp" />
+    <ClCompile Include="..\..\src\TEdge.cpp" />
     <ClCompile Include="..\..\src\TGraph.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -126,6 +127,7 @@
     <ClInclude Include="..\..\include\MyExceptions.h" />
     <ClInclude Include="..\..\include\TDHeap.h" />
     <ClInclude Include="..\..\include\TDisjointSet.h" />
+    <ClInclude Include="..\..\include\TEdge.h" />
     <ClInclude Include="..\..\include\TGraph.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
